refactoring and added examples
