+++ .github/workflows/ci-python-notebooks.yml
@@ -1,0 +1,44 @@
+# This is a basic workflow to help you get started with Actions
+
+name: CI-notebooks
+
+# Controls when the workflow will run
+on:
+  # Triggers the workflow on push or pull request events but only for the main branch
+  push:
+    branches: [ dev, release/* ]
+  pull_request:
+    branches: [ dev, release/* ]
+
+  # Allows you to run this workflow manually from the Actions tab
+  workflow_dispatch:
+
+# A workflow run is made up of one or more jobs that can run sequentially or in parallel
+jobs:
+  # This workflow contains a single job called "build"
+  build:
+    # The type of runner that the job will run on
+    runs-on:  ${{ matrix.platform }}
+    strategy:
+      max-parallel: 5
+      matrix:
+        platform: [ubuntu-latest]
+        python-version: [3.9]
+
+    # Steps represent a sequence of tasks that will be executed as part of the job
+    steps:
+      - uses: actions/checkout@v2
+      - name: Set up Python ${{ matrix.python-version }}
+        uses: actions/setup-python@v2.2.2
+        with:
+          python-version: ${{ matrix.python-version }}
+      - name: Install dependencies
+        run: |
+          python -m pip install --upgrade pip setuptools wheel
+          python -m pip install poetry
+          poetry install
+      - name: Generate python single files
+        run: |
+          ls examples | xargs -I {} bash -c "poetry run jupyter nbconvert --to python examples/{} --stdout > tests/examples/{}.py"
+      - name: Run examples
+        run: cd tests/examples && find . -name '*.py' -exec poetry run python {} \;
