+++ README.md
@@ -1,2 +1,34 @@
 # pams
 PAMS: Platform for Artificial Market Simulations
+
+[![python](https://img.shields.io/pypi/pyversions/pams.svg)](https://pypi.org/project/pams)
+[![pypi](https://img.shields.io/pypi/v/pams.svg)](https://pypi.org/project/pams)
+[![CI](https://github.com/masanorihirano/pams/actions/workflows/ci-python.yml/badge.svg)](https://github.com/masanorihirano/pams/actions/workflows/ci-python.yml)
+[![downloads](https://img.shields.io/pypi/dm/pams)](https://pypi.org/project/pams)
+[![code style: black](https://img.shields.io/badge/code%20style-black-000000.svg)](https://github.com/psf/black)
+[![Codacy Badge](https://app.codacy.com/project/badge/Grade/18ed1eecc4f34a99bb6fd9a7160f78ca)](https://www.codacy.com/gh/masanorihirano/pams/dashboard?utm_source=github.com&amp;utm_medium=referral&amp;utm_content=masanorihirano/pams&amp;utm_campaign=Badge_Grade)
+[![codecov](https://codecov.io/gh/masanorihirano/pams/branch/main/graph/badge.svg?token=tFccElw7Wd)](https://codecov.io/gh/masanorihirano/pams)
+
+## Documentations & User Guides
+
+Documentations are available on [readthedoc](https://pams.hirano.dev/)
+
+## Install
+This package is available on pypi as [`pams`](https://pypi.org/project/pams/)
+```bash
+$ pip install pams
+$ python
+>> import pams
+```
+
+
+## Issues and Contribution
+About issues (bugs):
+-   You can report issues [here](https://github.com/masanorihirano/slack_transfer/issues).
+-   日本語でissueを立てて構いません．
+-   There are no guarantee to support or fix those issues.
+
+Contributions:
+-   You can send pull requests (PRs) to this repository.
+-   But, there are no guarantee to merge your PRs.
+
