+++ pams/__init__.py
@@ -1,10 +1,10 @@
 from .fundamentals import Fundamentals
 from .index_market import IndexMarket
-from .logs import CancelLog
-from .logs import ExecutionLog
-from .logs import Log
-from .logs import Logger
-from .logs import OrderLog
+from .logs.base import CancelLog
+from .logs.base import ExecutionLog
+from .logs.base import Log
+from .logs.base import Logger
+from .logs.base import OrderLog
 from .market import Market
 from .order import LIMIT_ORDER
 from .order import MARKET_ORDER
