+++ pams/logs/base.py
@@ -194,10 +194,3 @@ class Logger:
 
     def process_market_step_end_log(self, log: "MarketStepEndLog") -> None:
         pass
-
-
-class MarketStepPrintLogger(Logger):
-    def process_market_step_end_log(self, log: MarketStepEndLog) -> None:
-        print(
-            f"{log.session.session_id} {log.market.get_time()} {log.market.market_id} {log.market.name} {log.market.get_market_price()} {log.market.get_fundamental_price()}"
-        )
