+++ pams/session.py
@@ -3,7 +3,7 @@ from typing import Any
 from typing import Dict
 from typing import Optional
 
-from .logs import Logger
+from .logs.base import Logger
 
 
 class Session:
