+++ pyproject.toml
@@ -21,6 +21,8 @@ pytest = "^7.1.3"
 pytest-cov = "^4.0.0"
 Sphinx = "^5.3.0"
 furo = "^2022.9.29"
+jupyter = "^1.0.0"
+matplotlib = "^3.6.2"
 
 [build-system]
 requires = ["poetry-core"]
