+++ tests/examples/.gitignore
@@ -1,0 +1,2 @@
+*
+!.gitignore
